// web/src/app/generated/model/book.ts
/**
 * Modul450 Library API
 * Simple library management system for renting books
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Category } from './category';


export interface Book { 
    /**
     * Unique identifier for the book
     */
    id: string;
    /**
     * Title of the book
     */
    name: string;
    /**
     * Author of the book
     */
    author: string;
    category: Category;
    /**
     * Whether the book is available for rent
     */
    available: boolean;
    /**
     * ID of the user who rented the book (null if available)
     */
    rentedBy?: string | null;
    /**
     * Date when the book was rented (null if available)
     */
    rentedDate?: string | null;
}